Split tests in two classes
